Double precission added for input and model object detection

+++ parking_project/api/models/netModel.py
@@ -236,8 +236,8 @@ class NetModel(models.Model):
     def detectOccupancyObjectDetection(self,image):
 
         self.model.eval()
-        self.model.to(DEVICE)
-        image = T.ToTensor()(image).to(DEVICE)
+        self.model.to(DEVICE).double()
+        image = T.ToTensor()(image).to(DEVICE).double()
 
         prediction = self.model([image])
